Add va_list types to bindgen whitelist, for cross-compilation
